fix: address migration linter review feedback
