+++ platform/backend/src/standalone-scripts/lint-drizzle-migrations.ts
@@ -1,16 +1,12 @@
-import { execFileSync } from "node:child_process";
 import path from "node:path";
 import {
+  findChangedMigrationFiles,
   type LintMigrationResult,
   lintMigrationFile,
   summarizeIssues,
 } from "@drizzle-migration-linter";
 
 const MIGRATIONS_DIR = path.resolve(process.cwd(), "src/database/migrations");
-const GIT_ROOT = execFileSync("git", ["rev-parse", "--show-toplevel"], {
-  encoding: "utf8",
-  cwd: process.cwd(),
-}).trim();
 
 function main(): void {
   const baseRef =
@@ -27,34 +23,11 @@ function main(): void {
 }
 
 function getChangedMigrationFiles(baseRef: string): string[] {
-  const resolvedBaseRef = ensureGitRefAvailable(baseRef);
-  const migrationsPathspec = path.relative(GIT_ROOT, MIGRATIONS_DIR);
-  const changedOutput = execFileSync(
-    "git",
-    [
-      "diff",
-      "--name-only",
-      "--diff-filter=ACMR",
-      resolvedBaseRef,
-      "--",
-      migrationsPathspec,
-    ],
-    { encoding: "utf8", cwd: GIT_ROOT },
-  );
-  const untrackedOutput = execFileSync(
-    "git",
-    ["ls-files", "--others", "--exclude-standard", "--", migrationsPathspec],
-    { encoding: "utf8", cwd: GIT_ROOT },
-  );
-
-  return `${changedOutput}\n${untrackedOutput}`
-    .split("\n")
-    .map((file) => file.trim())
-    .filter((file) => file.endsWith(".sql"))
-    .filter((file) => !file.includes("/meta/"))
-    .map((file) => path.resolve(GIT_ROOT, file))
-    .filter((file, index, files) => files.indexOf(file) === index)
-    .sort();
+  return findChangedMigrationFiles({
+    migrationsDir: MIGRATIONS_DIR,
+    baseRef,
+    options: { cwd: process.cwd() },
+  });
 }
 
 function printResults(
@@ -94,67 +67,6 @@ function printResults(
   process.stdout.write(
     `Drizzle migration linter found ${summary.errors} error${summary.errors === 1 ? "" : "s"} and ${summary.warnings} warning${summary.warnings === 1 ? "" : "s"}.\n`,
   );
-}
-
-function ensureGitRefAvailable(baseRef: string): string {
-  if (canResolveGitRef(baseRef)) {
-    return baseRef;
-  }
-
-  const remoteRef = parseRemoteRef(baseRef);
-  if (remoteRef) {
-    const { remote, branch } = remoteRef;
-    process.stderr.write(
-      `Drizzle migration linter base ref ${baseRef} is not available locally; fetching ${remote} ${branch}.\n`,
-    );
-    execFileSync(
-      "git",
-      [
-        "fetch",
-        "--depth=1",
-        remote,
-        `${branch}:refs/remotes/${remote}/${branch}`,
-      ],
-      { cwd: GIT_ROOT, stdio: "inherit" },
-    );
-    return baseRef;
-  }
-
-  const originRef = `origin/${baseRef}`;
-  process.stderr.write(
-    `Drizzle migration linter base ref ${baseRef} is not available locally; fetching origin ${baseRef}.\n`,
-  );
-  execFileSync(
-    "git",
-    [
-      "fetch",
-      "--depth=1",
-      "origin",
-      `${baseRef}:refs/remotes/origin/${baseRef}`,
-    ],
-    { cwd: GIT_ROOT, stdio: "inherit" },
-  );
-  return originRef;
-}
-
-function canResolveGitRef(ref: string): boolean {
-  try {
-    execFileSync("git", ["rev-parse", "--verify", `${ref}^{commit}`], {
-      cwd: GIT_ROOT,
-      stdio: "ignore",
-    });
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function parseRemoteRef(
-  ref: string,
-): { remote: string; branch: string } | null {
-  const [remote, ...branchParts] = ref.split("/");
-  if (!remote || branchParts.length === 0) return null;
-  return { remote, branch: branchParts.join("/") };
 }
 
 main();
